requirements.txt
pandas
requests
sentence-transformers
faiss-cpu
datasets
spellchecker
editdistance
nltk